requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.0.3

# Django Framework
django==4.2.7
django-admin-interface==0.30.0

# Data Processing
pandas==2.2.0
openpyxl==3.1.5

# Database
# sqlite3 is part of the standard library

# AI Integration
openai==1.10.0
tiktoken==0.5.0
langchain
langchain-openai

# Communication
# smtplib is part of the standard library

# Document Processing
python-docx
PyPDF2
chardet

# Utilities
python-dotenv==1.1.0
tenacity==8.5.0
regex==2023.12.25

# Testing
pytest==7.4.0
pytest-asyncio==0.23.0 